refactor: 금액 관련 필드를 Integer에서 BigDecimal로 변경
금융 도메인의 정확한 계산을 위해 모든 금액 관련 필드를 BigDecimal로 변경

주요 변경사항:
- Payment Entity: amount 필드 BigDecimal 변경 (precision=10, scale=2)
- PaymentRequest DTO: amount 필드 BigDecimal 변경
- PaymentService: processPayment() 메서드 파라미터 타입 변경
- OrderCreatedEvent: price/totalPrice 필드 BigDecimal 변경 및 productId 필드 추가
- InventoryReservedEvent: totalPrice 필드 BigDecimal 변경
- PaymentCompletedEvent: amount 필드 BigDecimal 변경
- Order Entity: productId 필드 추가, 생성자 수정
- OrderEventProducer: totalPrice 계산 로직 추가 (단가 × 수량)

+++ payment-service/src/main/java/com/example/payment/dto/PaymentRequest.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 /**
  * PG 결제 요청 DTO
  */
@@ -12,7 +14,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class PaymentRequest {
     private Long orderId;
-    private Integer amount;
+    private BigDecimal amount;
     private String paymentMethod;
     private String customerName;
     private String customerEmail;
